backend/pyproject.toml
[project]
name = "budget-backend"
version = "0.1.0"
description = "Budget App Backend with Django REST Framework"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
    "psycopg[binary]>=3.1.0",
    "django-cors-headers>=4.3.1",
    "python-dotenv>=1.0.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "dj-database-url>=3.0.1",
    "argon2-cffi>=23.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["budget_project", "api"]

